+++ ConsoleBalatro/Engine/Events/Args/EngineChipsMultGainEmitArgs.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -10,7 +11,7 @@ namespace ConsoleBalatro.Engine.Events.Args
     public class EngineChipsMultGainEmitArgs : EngineEventArgs
     {
         public Card SourceOfEmit;
-        public long ChipsGainEmitted = -1;
+        public BigInteger ChipsGainEmitted = -1;
         public double MultGainEmitted = -1;
         public double MultMultEmitted = -1;
     }
