+++ ConsoleBalatro/Engine/Events/Args/EngineGetBlindReqArgs.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,6 +9,6 @@ namespace ConsoleBalatro.Engine.Events.Args
 {
     public class EngineGetBlindReqArgs : EngineEventArgs
     {
-        public long ChipRequirementAmount { get; set; }
+        public BigInteger ChipRequirementAmount { get; set; }
     }
 }
