+++ ConsoleBalatro/Engine/Events/Args/EngineHandLevelChangeArgs.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -11,9 +12,9 @@ namespace ConsoleBalatro.Engine.Events.Args
         public PlayedHandType HandTypeLevelling { get; set; }
         public bool isLevelUp { get; set; }
 
-        public int oldChipAmount { get; set; }
-        public int newChipAmount { get; set; }
-        public int chipChangeAmount { get; set; }
+        public BigInteger oldChipAmount { get; set; }
+        public BigInteger newChipAmount { get; set; }
+        public BigInteger chipChangeAmount { get; set; }
 
         public double oldMultAmount { get; set; }
         public double newMultAmount { get; set; }
