+++ ConsoleBalatro/Engine/Events/Args/EnginePreFinalGainArgs.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,7 +9,7 @@ namespace ConsoleBalatro.Engine.Events.Args
 {
     public class EnginePreFinalGainArgs : EngineEventArgs
     {
-        public long FinalChips { get; set; }
+        public BigInteger FinalChips { get; set; }
         public double FinalMult { get; set; }
     }
 }
